// src/types/sign-in.types.ts
export enum ESignInMessage {
  OAuthSignin = 'OAuthSignin',
  OAuthCallback = 'OAuthCallback',
  OAuthCreateAccount = 'OAuthCreateAccount',
  EmailCreateAccount = 'EmailCreateAccount',
  Callback = 'Callback',
  OAuthAccountNotLinked = 'OAuthAccountNotLinked',
  EmailSignin = 'EmailSignin',
  CredentialsSignin = 'CredentialsSignin',
  SessionRequired = 'SessionRequired',
  Default = 'Default',
}